+++ scripts/copy-workspace.js
@@ -7,7 +7,9 @@ var currentPath = process.cwd() // should be node_modules/@dadi/web
 var workspacePath = path.join(currentPath, 'workspace')
 var destinationDir = path.join(currentPath, '../../../workspace')
 
-fs.copy(workspacePath, destinationDir, { overwrite: false }, (err) => {
-  if (err) return console.error(err)
-  console.log('Web workspace created at', destinationDir)
-})
+if (!fs.existsSync(destinationDir)) {
+  fs.copy(workspacePath, destinationDir, { overwrite: false }, (err) => {
+    if (err) return console.error(err)
+    console.log('Web workspace created at', destinationDir)
+  })
+}
